+++ examples/run_all_examples.py
@@ -30,6 +30,7 @@ EXAMPLE_FILES = [
     'analyses/gapfinder.py',
     'analyses/gridaggregator.py',
     'analyses/histogram.py',
+    'analyses/optimumrange.py',
 ]
 
 MAX_WORKERS = 4  # Number of parallel workers